+++ Makefile
@@ -1,7 +1,7 @@
 
 all: simulator
 
-simulator: main.cpp simulator.cpp cctree.cpp cctree.h heap.cpp heap.h
-	g++ -g -o simulator main.cpp simulator.cpp cctree.cpp heap.cpp
+simulator: main.cpp simulator.cpp cctree.cpp cctree.h ccnode.cpp ccnode.h heap.cpp heap.h
+	g++ -g -o simulator main.cpp simulator.cpp cctree.cpp ccnode.cpp heap.cpp
