+++ cctree.cpp
@@ -1,357 +1,177 @@
 #include "cctree.h"
+#include "method.h"
+int debug = 0;
 
-extern int thread_number;
-extern map<int, int> threadIdNumbering;
 
-MethodMap Method::allMethods;
+void CCTree::handle_object_allocation(int object_id, int size, std::string type, int thread_id, int method_id) {
 
-int CCNode::count = 0;
+  CCNode * curContext = 0;
 
-CCNode * CCNode::demand_child(int id, int thread_id, int time)
-{
-  for (CCNodeVector::iterator p = children.begin();
-    p != children.end();
-    p++)
-  {
-    CCNode * child = (*p);
-    if (child->method_id == id) {
-      child->setLastCall(time);
-      return child;
-    }
+  if (thread_id == object_id) {
+    // Spawning a new thread
+    // Make it a child of the root
+    curContext = root;
+    CCNode * newContext = curContext->demand_child(method_id,
+                                                   thread_id,
+                                                   time);
+    newContext->incCalls();
+    theStack[thread_id] = newContext;
+    if (debug > 0)
+      std::cout << "Spawn thread 0x" << std::hex << thread_id << std::dec << std::endl;
   }
 
-  // Not Found
-  CCNode * new_child = new CCNode(id, thread_id, time, this);
-  children.push_back(new_child);
-  return new_child;
-}
-
-void CCNode::computeTotals()
-{
-  // Start with this node's values
-  incTotalAllocBytes(getAllocBytes());
-  incTotalAllocObjects(getAllocObjects());
-  incTotalDeadBytes(getDeadBytes());
-  incTotalDeadObjects(getDeadObjects());
-
-  // Compute totals for any children, then add up
-  const CCNodeVector & children = getChildren();
-  for (CCNodeVector::const_iterator p = children.begin();
-       p != children.end();
-       p++) {
-      CCNode * child = (*p);
-      child->computeTotals();
-
-      incTotalAllocBytes(child->getTotalAllocBytes());
-      incTotalAllocObjects(child->getTotalAllocObjects());
-      incTotalDeadBytes(child->getTotalDeadBytes());
-      incTotalDeadObjects(child->getTotalDeadObjects());
+  if (theStack[thread_id] == 0) {
+    theStack[thread_id] = root;
   }
-}
 
-string CCNode::getMethodFullName()
-{
-  Method * meth = Method::getMethod(getMethodId());
-  string result;
-  if (meth)
-    result = meth->getClassName() + "::" + meth->getMethodName();
-  else
-    result = "ENTRY";
+  curContext = theStack[thread_id];
 
-  return result;
-}
+  curContext->incAllocBytes(size);
+  curContext->incAllocObjects();
+  heapObject = HeapObject::DemandHeapObject(object_id);
+  heapObject->setAlloc(time, size, type);
+  heapObject->setAllocCC(curContext);
 
-void CCNode::printTree(int depth)
-{
-  if (getTotalAllocBytes() == 0 &&
-      getTotalAllocObjects() == 0)
-    return;
+  if (debug > 0) {
+    std::cout << "Allocate 0x" << std::hex << object_id << " size 0x" << size
+         << std::dec << " at time " << time << std::endl;
 
-  for (int i = 0; i < depth; i++)
-    cout << ".";
-
-  if (depth > 0)
-    cout << " ";
-
-  cout << getMethodFullName() << " ";
-
-  cout << getTotalAllocBytes() << " "
-       << getAllocRank() << " "
-       << getTotalAllocObjects() << "  "
-       << getCalls() << endl;
-
-  const CCNodeVector & children = getChildren();
-  for (CCNodeVector::const_iterator p = children.begin();
-       p != children.end();
-       p++)
-    {
-      CCNode * child = (*p);
-      child->printTree(depth+1);
-    }
-}
-
-void CCNode::collectNodes(CCNodeVector & all)
-{
-  all.push_back(this);
-  const CCNodeVector & children = getChildren();
-  for (CCNodeVector::const_iterator p = children.begin();
-       p != children.end();
-       p++)
-    {
-      CCNode * child = (*p);
-      child->collectNodes(all);
-    }
-}
-
-bool compareNodes(CCNode * one, CCNode * two)
-{
-  return one->getTotalAllocBytes() < two->getTotalAllocBytes();
-}
-
-void CCNode::rankNodes()
-{
-  CCNodeVector all;
-  collectNodes(all);
-  sort(all.begin(), all.end(), compareNodes);
-
-  int rank = all.size();
-  for (CCNodeVector::const_iterator p = all.begin();
-       p != all.end();
-       p++)
-    {
-      CCNode * child = (*p);
-      child->setAllocRank(rank);
-      rank--;
-    }
-}
-
-void CCNode::printStack()
-{
-  CCNode * cur = this;
-  while (cur) {
-    cout << "  " << cur->getMethodFullName()
-         << "(0x" << hex << cur->getMethodId()
-         << dec << ")" << endl;
-
-    cur = cur->getParent();
+    //if (debug > 1) curContext->printStack();
   }
+  last_thread_id = thread_id;
 }
 
-void CCNode::emitPath(ofstream & out)
-{
-  CCNode * parent = getParent();
-  if (parent != 0)
-    parent->emitPath(out);
+void CCTree::handle_object_death(int object_id) {
+  heapObject = HeapObject::DemandHeapObject(object_id);
+  heapObject->setDead(time);
+  (HeapObject::Find(heapObject))->incNumDead();
 
-  out << "\t" << getMethodFullName();
+  CCNode * curContext = theStack[last_thread_id];
+
+  curContext->incDeadBytes(heapObject->getSize());
+  curContext->incDeadObjects();
+  heapObject->setDeathCC(curContext);
 }
 
-void CCNode::emitTreeMapTM3Rec(ofstream & out)
-{
-  int abytes = getTotalAllocBytes();
-  int dbytes = getTotalDeadBytes();
-
-  if (abytes > 0 || dbytes > 0) {
-    const CCNodeVector & children = getChildren();
-
-    // Compute total size of children
-    int child_total_alloc = 0;
-    int child_total_dead = 0;
-    int num_children = 0;
-    for (CCNodeVector::const_iterator p = children.begin();
-      p != children.end();
-      p++)
-    {
-      CCNode * child = (*p);
-      int child_alloc = child->getTotalAllocBytes();
-      int child_dead =  child->getTotalDeadBytes();
-      child_total_alloc += child_alloc;
-      child_total_dead  += child_dead;
-      if (child_alloc > 0 || child_dead > 0)
-        num_children++;
-    }
-
-    // Output entry for this node
-    out << (abytes+dbytes) << "\t";
-    out << getCalls() << "\t";
-    out << "0" << "\t";
-    out << getThreadId() << "\t";
-    out << getFirstCall() << "\t";
-    emitPath(out);
-    out << endl;
-
-    // Output synthetic node representing the this node's contribution
-    // (if necessary)
-    if (num_children > 0) {
-      int alloc_diff = abytes - child_total_alloc;
-      int dead_diff = dbytes - child_total_dead;
-      if (alloc_diff > 0) {
-        out << alloc_diff << "\t";
-        out << getCalls() << "\t";
-        out << "0" << "\t";
-        out << getThreadId() << "\t";
-        out << getFirstCall() << "\t";
-        emitPath(out);
-        out << "\talloc";
-        out << endl;
+void CCTree::handle_object_update(int old_target, int object_id, int new_target, int target_id) {
+  if (object_id != 0 && new_target != 0) {
+    heapObject = HeapObject::DemandHeapObject(object_id);
+    targetObject = HeapObject::DemandHeapObject(new_target);
+    HeapObject * sroot = HeapObject::Find(heapObject);
+    HeapObject * troot = HeapObject::Find(targetObject);
+    if (targetObject->getType() != "[C" &&
+        targetObject->getType() != "UNKNOWN" &&
+        troot->getId() != 0x19a) {
+      if (debug > 0) {
+        std::cout << "Pointer from 0x" << std::hex << object_id << " "
+             << heapObject->getType() << " to 0x" << new_target
+                  << std::dec << " " << targetObject->getType() << " at time "
+             << time << std::endl;
+        // if (debug > 1) {
+        //   CCNode * curContext = theStack[last_thread_id];
+        //   curContext->printStack();
+        // }
       }
-      if (dead_diff > 0) {
-        out << dead_diff << "\t";
-        out << getCalls() << "\t";
-        out << "0" << "\t";
-        out << getThreadId() << "\t";
-        out << getFirstCall() << "\t";
-        emitPath(out);
-        out << "\tdead";
-        out << endl;
-      }
-    }
-    // Output the children
-    for (CCNodeVector::const_iterator p = children.begin();
-      p != children.end();
-      p++)
-    {
-      CCNode * child = (*p);
-      child->emitTreeMapTM3Rec(out);
+
+      // HeapObject::Union(heapObject, targetObject);
+      heapObject->setPointsTo(targetObject);
     }
   }
 }
 
-void CCNode::emitTreeMapTM3(ofstream & out)
-{
-  out << "Bytes\tCalls\tLifetime\tThread\tFirst call" << endl;
-  out << "INTEGER\tINTEGER\tINTEGER\tINTEGER\tINTEGER" << endl;
-
-  emitTreeMapTM3Rec(out);
-}
-
-// Emit the calling context tree in TreeML format, for use in prefuse
-void CCNode::emitTreeML(ofstream & out)
-{
-  out << "<tree>" << endl;
-  out << " <declarations>" << endl;
-  out << "  <attributeDecl name=\"name\" type=\"String\"/>" << endl;
-  out << "  <attributeDecl name=\"class\" type=\"String\"/>" << endl;
-  out << "  <attributeDecl name=\"bytes\" type=\"Long\"/>" << endl;
-  out << "  <attributeDecl name=\"alloc\" type=\"Long\"/>" << endl;
-  out << "  <attributeDecl name=\"dead\" type=\"Long\"/>" << endl;
-  out << "  <attributeDecl name=\"objects\" type=\"Long\"/>" << endl;
-  out << "  <attributeDecl name=\"objalloc\" type=\"Long\"/>" << endl;
-  out << "  <attributeDecl name=\"objdead\" type=\"Long\"/>" << endl;
-  out << "  <attributeDecl name=\"thread\" type=\"Long\"/>" << endl;
-  out << " </declarations>" << endl;
-
-  emitTreeMLRec(out, 1);
-
-  out << "</tree>" << endl;
-}
-
-void CCNode::emitTreeMLRec(ofstream & out, int depth)
-{
-  int abytes = getTotalAllocBytes();
-  int dbytes = getTotalDeadBytes();
-
-  if (abytes + dbytes > 0) {
-    const CCNodeVector & children = getChildren();
-    bool is_leaf = (children.size() == 0);
-
-    if (is_leaf) {
-      // -- Leaf
-      out << "<leaf>" << endl;
+void CCTree::handle_method_entry(int method_id, int object_id, int thread_id) {
+  CCNode * curContext = theStack[thread_id];
+  bool new_thread = false;
+  if (curContext == 0) {
+    // Spawning a new thread -- look up where the thread started.
+    // Relies on the fact that the thread_id is the same as the
+    // object_id of the Thread object instance.
+    new_thread = true;
+    curContext = threadStarts[thread_id];
+    if (curContext) {
+      std::cout << "Spawn thread 0x" << std::hex << thread_id << std::dec
+           << " -- started at " << curContext->getMethodFullName()
+           << std::endl;
+      std::cout << "   in context" << std::endl;
+      //curContext->printStack();
     } else {
-      out << "<branch>" << endl;
-    }
-
-    Method * meth = Method::getMethod(getMethodId());
-    string nm;
-    string classnm;
-    if (meth) {
-      nm = meth->getMethodName();
-      classnm = meth->getClassName();
-    } else {
-      nm = "ENTRY";
-      classnm = "";
-    }
-
-    int i = nm.find("<init>");
-    if (i != string::npos)
-      nm.replace(i, i+6, "&lt;init&gt;");
-    int j = nm.find("<clinit>");
-    if (j != string::npos)
-      nm.replace(j, j+8, "&lt;clinit&gt;");
-    out << "<attribute name=\"name\" value=\"" << nm << "\"/>" << endl;
-    out << "<attribute name=\"class\" value=\"" << classnm << "\"/>" << endl;
-    out << "<attribute name=\"bytes\" value=\"" << (abytes+dbytes)
-        << "\"/>" << endl;
-    out << "<attribute name=\"alloc\" value=\"" << abytes << "\"/>" << endl;
-    out << "<attribute name=\"dead\" value=\"" << dbytes << "\"/>" << endl;
-    out << "<attribute name=\"objects\" value=\""
-        << (getAllocObjects() + getDeadObjects()) << "\"/>" << endl;
-    out << "<attribute name=\"objalloc\" value=\"" << getAllocObjects()
-        << "\"/>" << endl;
-    out << "<attribute name=\"objdead\" value=\"" << getDeadObjects()
-        << "\"/>" << endl;
-    out << "<attribute name=\"thread\" value=\""
-        << threadIdNumbering[getThreadId()] << "\"/>" << endl;
-
-    for (CCNodeVector::const_iterator p = children.begin();
-      p != children.end();
-      p++)
-    {
-      CCNode * child = (*p);
-      child->emitTreeMLRec(out, depth+1);
-    }
-
-    if (is_leaf) {
-      // -- Leaf
-      out << "</leaf>" << endl;
-    } else {
-      out << "</branch>" << endl;
+      std::cout << "Problem: no threadStart for thread id 0x" << std::hex
+                << thread_id << std::dec << std::endl;
+      curContext = root;
     }
   }
-}
 
-void CCNode::emitTreeJSON(ofstream & out)
-{
-  emitTreeJSONRec(out, 0);
-}
+  time++;
+  depth++;
 
-void CCNode::emitTreeJSONRec(ofstream & out, int depth)
-{
-  int abytes = getTotalAllocBytes();
-  int dbytes = getTotalDeadBytes();
+  curContext = curContext->demand_child(method_id, thread_id, time);
+  curContext->incCalls();
+  theStack[thread_id] = curContext;
 
-  if (abytes + dbytes > 0) {
-    string space(depth*2, ' ');
-    out << space << "{" << endl;
-
-    Method * meth = Method::getMethod(getMethodId());
-    string nm;
-    string classnm;
-    if (meth) {
-      nm = meth->getMethodName();
-      classnm = meth->getClassName();
-    } else {
-      nm = "ENTRY";
-      classnm = "";
-    }
-
-    out << space << " \"name\": \"" << nm << "\"," << endl;
-
-    const CCNodeVector & children = getChildren();
-    if (children.size() == 0) {
-      out << space << " \"size\": " << abytes << endl;
-    } else {
-      out << space << " \"children\": [" << abytes << endl;
-      for (CCNodeVector::const_iterator p = children.begin();
-           p != children.end();
-           p++)
-      {
-        CCNode * child = (*p);
-        child->emitTreeJSONRec(out, depth+1);
-      }
-      cout << space << " ]" << endl;
-    }
-    out << space << "}" << endl;
+  if (debug > 0 || new_thread) {
+    std::cout << "Enter " << curContext->getMethodFullName() << " 0x"
+              << std::hex << method_id << " thread 0x" << thread_id << std::dec
+         << " at time " << time << std::endl;
+    //if (debug > 1) curContext->printStack();
   }
+
+  if (method_id == thread_start_method_id) {
+    // -- Found a new thread start
+    threadStarts[object_id] = curContext;
+    thread_number++;
+    threadIdNumbering[object_id] = thread_number;
+    if (true) {
+      std::cout << "Found Thread.start at " << std::endl;
+      //curContext->printStack();
+    }
+  }
+
+  last_thread_id = thread_id;
+}
+
+void CCTree::handle_method_exit(int method_id, int object_id, int thread_id) {
+  CCNode * curContext = theStack[thread_id];
+  if (theStack.find(thread_id) == theStack.end()) {
+    std::cout << "Never seen thread 0x" << std::hex << thread_id << std::endl;
+  }
+
+  if (curContext == 0) {
+    Method * meth = Method::getMethod(method_id);
+    std::cout << "ERROR: Thread 0x" << std::hex << thread_id
+         << " returning from method " << meth->getMethodName()
+         << " in class " << meth->getClassName() << std::endl;
+    exit(EXIT_FAILURE);
+  }
+
+  if (debug > 0) {
+    std::cout << "Exit  " << curContext->getMethodFullName() << " 0x"
+              << std::hex << method_id << " thread 0x" << thread_id << std::dec
+         << " at time " << time << std::endl;
+    //if (debug > 1) curContext->printStack();
+  }
+
+  time++;
+
+  CCNode * returning = curContext;
+  int cur_id = returning->getMethodId();
+  int orig_depth = depth;
+  while (returning || returning->getMethodId() != method_id) {
+    returning = returning->getParent();
+    depth--;
+  }
+
+  if (returning == 0) {
+    std::cout << "THIS IS BAD: looking for "
+         << std::hex << "0x" << method_id << " but found 0x"
+              << cur_id << std::dec << std::endl;
+    //returning->printStack();
+    // current_node unchanged
+    depth = orig_depth;
+  } else {
+    returning->setLastCall(time);
+    theStack[thread_id] = returning->getParent();
+    depth--;
+  }
+
+  last_thread_id = thread_id;
 }
